rungen: moving extra functions to the "helpers" submodule
